requirements.txt
django
djangorestframework
factory-boy
faker
ply
psycopg2-binary
pygments
tqdm
